+++ lib/src/view/platforms/run_webapp.dart
@@ -41,7 +41,7 @@ void runApp(
   // Instantiate the app's error handler.
   final handler = v.AppErrorHandler(
       handler: errorHandler,
-      builder: errorScreen,
+      screen: errorScreen,
       report: errorReport,
       allowNewHandlers: allowNewHandlers);