+++ lib/src/view/utils/error_handler.dart
@@ -46,6 +46,7 @@ import 'package:flutter/rendering.dart'
         Paint,
         PaintingContext,
         RenderBox,
+        RenderObjectWithChildMixin,
         Size,
         StringProperty,
         TextAlign,
@@ -61,14 +62,15 @@ class AppErrorHandler {
   /// Optionally supply the Error handler, Builder, and Report routines.
   factory AppErrorHandler({
     FlutterExceptionHandler? handler,
-    ErrorWidgetBuilder? builder,
+    ErrorWidgetBuilder? screen,
     ReportErrorHandler? report,
     bool? allowNewHandlers,
   }) {
-    _this ??= AppErrorHandler._(builder);
+    //
+    _this ??= AppErrorHandler._();
 
     /// Allows you to set an error handler more than once.
-    final reassigned = set(handler: handler, builder: builder, report: report);
+    final reassigned = set(handler: handler, screen: screen, report: report);
 
     // Allow for null. Simply allow new handles by default
     allowNewHandlers ??= true;
@@ -81,7 +83,7 @@ class AppErrorHandler {
     return _this!;
   }
 
-  AppErrorHandler._(ErrorWidgetBuilder? builder) {
+  AppErrorHandler._() {
     // Record the current error handler.
     _oldOnError = FlutterError.onError;
 
@@ -89,9 +91,7 @@ class AppErrorHandler {
     _oldBuilder = ErrorWidget.builder;
 
     // At the start, define our own 'error building widget' widget if one is not provided.
-    builder ??= errorDisplayWidget;
-
-    ErrorWidget.builder = builder;
+    ErrorWidget.builder = errorDisplayWidget;
 
     FlutterError.onError = (FlutterErrorDetails details) {
       // Prevent an infinite loop and fall back to the original handler.
@@ -151,7 +151,7 @@ class AppErrorHandler {
   /// Set a handler and the report
   static bool set({
     FlutterExceptionHandler? handler,
-    ErrorWidgetBuilder? builder,
+    ErrorWidgetBuilder? screen,
     ReportErrorHandler? report,
   }) {
     // Once you're not allowed to reset the handlers, it can't be reversed.
@@ -174,9 +174,9 @@ class AppErrorHandler {
       reset = true;
     }
 
-    if (builder != null) {
+    if (screen != null) {
       // Change the widget presented when another widget fails to build.
-      ErrorWidget.builder = builder;
+      ErrorWidget.builder = screen;
       reset = true;
     }
     // Something was set;
@@ -299,80 +299,281 @@ class AppErrorHandler {
     } catch (e) {
       message = 'Error';
     }
-
-    final exception = details.exception;
-    return DisplayErrorWidget(
-        message: message, error: exception is Error ? exception : null);
+    return _ErrorRenderObjectWidget(message: message, error: details.exception);
   }
 }
 
 /// A low-level widget to present instead of the failed widget.
-class DisplayErrorWidget extends LeafRenderObjectWidget {
+class _ErrorRenderObjectWidget extends LeafRenderObjectWidget {
   /// Supply an error message to display and or a Error object.
-  DisplayErrorWidget({this.message = '', Error? error})
-      : _error = error,
-        super(key: UniqueKey());
+  const _ErrorRenderObjectWidget({
+    super.key,
+    this.message,
+    this.error,
+    this.paragraphStyle,
+    this.textStyle,
+    this.padding,
+    this.minimumWidth,
+    this.backgroundColor,
+  });
 
   /// The message to display.
-  final String message;
-  final Error? _error;
+  final String? message;
+
+  ///
+  final Object? error;
+
+  ///
+  final ui.ParagraphStyle? paragraphStyle;
+
+  ///
+  final ui.TextStyle? textStyle;
+
+  ///
+  final EdgeInsets? padding;
+
+  ///
+  final double? minimumWidth;
+
+  ///
+  final Color? backgroundColor;
 
   @override
-  RenderBox createRenderObject(BuildContext context) =>
-      _ErrorBox(_errorMessage());
+  RenderBox createRenderObject(BuildContext context) => _ErrorBox(
+        message: _errorMessage(message, error),
+        paragraphStyle: paragraphStyle,
+        textStyle: textStyle,
+        padding: padding,
+        minimumWidth: minimumWidth,
+        backgroundColor: backgroundColor,
+      );
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
-    if (_error == null || _error is! FlutterError) {
-      properties.add(StringProperty('message', _errorMessage(), quoted: false));
+    if (error == null || error is! FlutterError) {
+      properties.add(StringProperty('message', _errorMessage(message, error),
+          quoted: false));
     } else {
-      final FlutterError _flutterError = _error as FlutterError;
+      final FlutterError _flutterError = error as FlutterError;
       properties.add(_flutterError.toDiagnosticsNode(
           style: DiagnosticsTreeStyle.whitespace));
     }
   }
+}
 
-  // Compose an error message to be displayed.
-  // An empty string if no message was provided.
-  String _errorMessage() {
-    String _message;
-    if (message.isEmpty) {
-      if (_error == null) {
-        _message = '';
-      } else {
-        _message = _error.toString();
+///
+class ErrorWidgetDisplay {
+  ///
+  const ErrorWidgetDisplay({
+    this.paragraphStyle,
+    this.textStyle,
+    this.padding,
+    this.minimumWidth,
+    this.backgroundColor,
+    this.stackTrace = false,
+  });
+
+  ///
+  final ui.ParagraphStyle? paragraphStyle;
+
+  ///
+  final ui.TextStyle? textStyle;
+
+  ///
+  final EdgeInsets? padding;
+
+  ///
+  final double? minimumWidth;
+
+  ///
+  final Color? backgroundColor;
+
+  ///
+  final bool? stackTrace;
+
+  ///
+  Widget builder(FlutterErrorDetails details) {
+    String? message;
+    try {
+      //
+      if (stackTrace != null && stackTrace!) {
+        //
+        message = 'ERROR\n\n${details.exception}\n\n';
+
+        final stack = details.stack.toString().split('\n');
+
+        final length = stack.length > 5 ? 5 : stack.length;
+
+        final buffer = StringBuffer()..write(message);
+
+        for (var i = 0; i < length; i++) {
+          buffer.write('${stack[i]}\n');
+        }
+        message = buffer.toString();
       }
-    } else {
-      _message = message;
+    } catch (e) {
+      message = null;
     }
-    return _message;
+    return errorWidgetChild(
+      details: details,
+      message: message,
+      paragraphStyle: paragraphStyle,
+      textStyle: textStyle,
+      padding: padding,
+      minimumWidth: minimumWidth,
+      backgroundColor: backgroundColor,
+    );
+  }
+
+  ///
+  Widget errorWidgetChild({
+    Key? key,
+    FlutterErrorDetails? details,
+    String? message,
+    Object? error,
+    ui.ParagraphStyle? paragraphStyle,
+    ui.TextStyle? textStyle,
+    EdgeInsets? padding,
+    double? minimumWidth,
+    Color? backgroundColor,
+    CustomPainter? customPainter,
+  }) {
+    //
+    if (details != null) {
+      //
+      error ??= details.exception;
+    }
+
+    // Default
+    customPainter ??= _ErrorSymbol();
+
+    //
+    const double WIDTH = 30;
+
+    return ErrorRenderObjectWidget(
+      key: key,
+      message: message,
+      error: error,
+      paragraphStyle: paragraphStyle,
+      textStyle: textStyle,
+      padding: padding,
+      minimumWidth: minimumWidth,
+      backgroundColor: backgroundColor,
+      child: CustomPaint(
+        size: Size(WIDTH, (WIDTH * 1).toDouble()),
+        painter: customPainter,
+      ),
+    );
   }
 }
 
-class _ErrorBox extends RenderBox {
+///
+class ErrorRenderObjectWidget extends SingleChildRenderObjectWidget {
+  ///
+  const ErrorRenderObjectWidget({
+    super.key,
+    this.message,
+    this.error,
+    this.paragraphStyle,
+    this.textStyle,
+    this.padding,
+    this.minimumWidth,
+    this.backgroundColor,
+    super.child,
+  });
+
+  ///
+  final Object? error;
+
+  /// The message to display.
+  final String? message;
+
+  ///
+  final ui.ParagraphStyle? paragraphStyle;
+
+  ///
+  final ui.TextStyle? textStyle;
+
+  ///
+  final EdgeInsets? padding;
+
+  ///
+  final double? minimumWidth;
+
+  ///
+  final Color? backgroundColor;
+
+  @override
+  RenderBox createRenderObject(BuildContext context) => _ErrorBox(
+        message: _errorMessage(message, error),
+        paragraphStyle: paragraphStyle,
+        textStyle: textStyle,
+        padding: padding,
+        minimumWidth: minimumWidth,
+        backgroundColor: backgroundColor,
+      );
+}
+
+/// Compose an error message to be displayed.
+/// An empty string if no message was provided.
+String _errorMessage(String? message, Object? error) {
+  String _message;
+  if (message == null || message.isEmpty) {
+    if (error == null) {
+      _message = '';
+    } else {
+      _message = error.toString();
+    }
+  } else {
+    _message = message;
+  }
+  return _message;
+}
+
+class _ErrorBox extends RenderBox with RenderObjectWithChildMixin<RenderBox> {
   ///
   /// A message can optionally be provided. If a message is provided, an attempt
   /// will be made to render the message when the box paints.
-  _ErrorBox([this.message = '']) {
+  _ErrorBox({
+    this.message = '',
+    ui.ParagraphStyle? paragraphStyle,
+    ui.TextStyle? textStyle,
+    EdgeInsets? padding,
+    double? minimumWidth,
+    Color? backgroundColor,
+  }) {
+    //
+    if (paragraphStyle != null) {
+      _paragraphStyle = paragraphStyle;
+    }
+    _textStyle = textStyle ?? _initTextStyle();
+    if (padding != null) {
+      _padding = padding;
+    }
+    if (minimumWidth != null) {
+      _minimumWidth = minimumWidth;
+    }
+    _backgroundColor = backgroundColor ?? _initBackgroundColor();
+    final _message = message == null || message!.isEmpty
+        ? 'Unknown Error!'
+        : message!.trim();
     try {
-      if (message != '') {
-        ///
-        /// Generally, the much better way to draw text in a RenderObject is to
-        /// use the TextPainter class. If you're looking for code to crib from,
-        /// see the paragraph.dart file and the RenderParagraph class.
-        final builder = ui.ParagraphBuilder(paragraphStyle)
-          ..pushStyle(textStyle)
-          ..addText(message);
-        _paragraph = builder.build();
-      }
+      ///
+      /// Generally, the much better way to draw text in a RenderObject is to
+      /// use the TextPainter class. If you're looking for code to crib from,
+      /// see the paragraph.dart file and the RenderParagraph class.
+      final builder = ui.ParagraphBuilder(_paragraphStyle)
+        ..pushStyle(_textStyle)
+        ..addText(_message);
+      _paragraph = builder.build();
     } catch (error) {
       // Intentionally left empty.
     }
   }
 
   /// The message to attempt to display at paint time.
-  final String message;
+  final String? message;
 
   ui.Paragraph? _paragraph;
 
@@ -407,22 +608,22 @@ class _ErrorBox extends RenderBox {
   ///
   /// See also:
   ///
-  ///  * [minimumWidth], which controls how wide the box must be before the
+  ///  * [_minimumWidth], which controls how wide the box must be before the
   //     horizontal padding is applied.
-  static EdgeInsets padding = const EdgeInsets.fromLTRB(34, 96, 34, 12);
+  EdgeInsets _padding = const EdgeInsets.fromLTRB(34, 96, 34, 12);
 
   /// The width below which the horizontal padding is not applied.
   ///
   /// If the left and right padding would reduce the available width to less than
   /// this value, then the text is rendered flush with the left edge.
-  static double minimumWidth = 200;
+  double _minimumWidth = 200;
 
   /// The color to use when painting the background of [RenderBox] objects.
   /// a red from a light gray.
-  static Color backgroundColor = _initBackgroundColor();
+  late Color _backgroundColor;
 
-  /// Ligt gray in production; Red in development.
-  static Color _initBackgroundColor() {
+  /// Light gray in production; Red in development.
+  Color _initBackgroundColor() {
     var result = const Color(0xF0C0C0C0);
     assert(() {
       result = const Color(0xF0900000);
@@ -433,10 +634,10 @@ class _ErrorBox extends RenderBox {
 
   /// The text style to use when painting [RenderBox] objects.
   /// a dark gray sans-serif font.
-  static ui.TextStyle textStyle = _initTextStyle();
+  late ui.TextStyle _textStyle;
 
   /// Black text in production; Yellow in development.
-  static ui.TextStyle _initTextStyle() {
+  ui.TextStyle _initTextStyle() {
     var result = ui.TextStyle(
       color: const Color(0xFF303030),
       fontFamily: 'sans-serif',
@@ -455,7 +656,7 @@ class _ErrorBox extends RenderBox {
   }
 
   /// The paragraph style to use when painting [RenderBox] objects.
-  static ui.ParagraphStyle paragraphStyle = ui.ParagraphStyle(
+  ui.ParagraphStyle _paragraphStyle = ui.ParagraphStyle(
     textDirection: TextDirection.ltr,
     textAlign: TextAlign.left,
   );
@@ -463,23 +664,339 @@ class _ErrorBox extends RenderBox {
   @override
   void paint(PaintingContext context, Offset offset) {
     try {
-      context.canvas.drawRect(offset & size, Paint()..color = backgroundColor);
+      context.canvas.drawRect(offset & size, Paint()..color = _backgroundColor);
       if (_paragraph != null) {
         var width = size.width;
         var left = 0.0;
         var top = 0.0;
-        if (width > padding.left + minimumWidth + padding.right) {
-          width -= padding.left + padding.right;
-          left += padding.left;
+        if (width > _padding.left + _minimumWidth + _padding.right) {
+          width -= _padding.left + _padding.right;
+          left += _padding.left;
         }
         _paragraph!.layout(ui.ParagraphConstraints(width: width));
-        if (size.height > padding.top + _paragraph!.height + padding.bottom) {
-          top += padding.top;
+        if (size.height > _padding.top + _paragraph!.height + _padding.bottom) {
+          top += _padding.top;
         }
         context.canvas.drawParagraph(_paragraph!, offset + Offset(left, top));
       }
     } catch (ex) {
       // Intentionally left empty.
     }
+  }
+}
+
+/// The Error symbol programmatically displayed.
+class _ErrorSymbol extends CustomPainter {
+  ///
+  // ignore: non_constant_identifier_names
+  final NaN = double.nan;
+
+  //
+  @override
+  void paint(Canvas canvas, Size size) {
+    final Path path_0 = Path();
+    path_0.moveTo(size.width * 4.414063, size.height * 9.287109);
+    path_0.cubicTo(
+        size.width * 2.734375,
+        size.height * 9.044922,
+        size.width * 1.376953,
+        size.height * 7.898438,
+        size.width * 0.8710938,
+        size.height * 6.291016);
+    path_0.cubicTo(
+        size.width * 0.7382813,
+        size.height * 5.873047,
+        size.width * 0.6835938,
+        size.height * 5.494141,
+        size.width * 0.6835938,
+        size.height * 5.000000);
+    path_0.cubicTo(
+        size.width * 0.6835938,
+        size.height * 4.400391,
+        size.width * 0.7714844,
+        size.height * 3.921875,
+        size.width * 0.9785156,
+        size.height * 3.404297);
+    path_0.cubicTo(
+        size.width * 1.037109,
+        size.height * 3.255859,
+        size.width * 1.035156,
+        size.height * 3.187500,
+        size.width * 0.9687500,
+        size.height * 3.150391);
+    path_0.cubicTo(
+        size.width * 0.9453125,
+        size.height * 3.136719,
+        size.width * 0.9570313,
+        size.height * 3.103516,
+        size.width * 1.031250,
+        size.height * 2.958984);
+    path_0.cubicTo(
+        size.width * 1.119141,
+        size.height * 2.792969,
+        size.width * 1.126953,
+        size.height * 2.783203,
+        size.width * 1.181641,
+        size.height * 2.789063);
+    path_0.cubicTo(
+        size.width * 1.259766,
+        size.height * 2.796875,
+        size.width * 1.283203,
+        size.height * 2.775391,
+        size.width * 1.423828,
+        size.height * 2.564453);
+    path_0.cubicTo(
+        size.width * 2.054688,
+        size.height * 1.615234,
+        size.width * 3.083984,
+        size.height * 0.9492188,
+        size.width * 4.248047,
+        size.height * 0.7363281);
+    path_0.cubicTo(
+        size.width * 4.451172,
+        size.height * 0.6992188,
+        size.width * 4.205078,
+        size.height * 1.488281,
+        size.width * 4.205078,
+        size.height * 1.294922);
+    path_0.cubicTo(
+        size.width * 4.207031,
+        size.height * 1.964844,
+        size.width * 4.218750,
+        size.height * 2.400391,
+        size.width * 4.332031,
+        size.height * 2.220703);
+    path_0.cubicTo(
+        size.width * 4.570313,
+        size.height * 4.015625,
+        size.width * 5.158203,
+        size.height * 4.582031,
+        size.width * 5.923828,
+        size.height * 2.623047);
+    path_0.cubicTo(
+        size.width * 6.468750,
+        size.height * 3.283203,
+        size.width * 7.064453,
+        size.height * 4.458984,
+        size.width * 6.001953,
+        size.height * 3.023438);
+    path_0.cubicTo(
+        size.width * 6.070313,
+        size.height * 4.304688,
+        size.width * 5.988281,
+        size.height * 4.609375,
+        size.width * 5.992188,
+        size.height * 3.056641);
+    path_0.cubicTo(
+        size.width * 5.994141,
+        size.height * 3.115234,
+        size.width * 6.041016,
+        size.height * 3.158203,
+        size.width * NaN,
+        size.height * NaN);
+    path_0.lineTo(size.width * NaN, size.height * NaN);
+    path_0.lineTo(size.width * NaN, size.height * NaN);
+    path_0.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_0.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_0.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_0.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_0.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_0.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_0.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_0.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_0.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_0.close();
+    path_0.moveTo(size.width * 5.314453, size.height * 8.359375);
+    path_0.cubicTo(
+        size.width * 6.480469,
+        size.height * 8.240234,
+        size.width * 7.494141,
+        size.height * 7.550781,
+        size.width * 3.478516,
+        size.height * 9.015625);
+    path_0.cubicTo(
+        size.width * 2.166016,
+        size.height * 9.423828,
+        size.width * 0.6035156,
+        size.height * 8.390625,
+        size.width * -0.4277344,
+        size.height * 7.759766);
+    path_0.cubicTo(
+        size.width * -1.683594,
+        size.height * 4.501953,
+        size.width * -1.750000,
+        size.height * 3.150391,
+        size.width * -0.5800781,
+        size.height * 7.453125);
+    path_0.cubicTo(
+        size.width * -1.173828,
+        size.height * 8.089844,
+        size.width * -1.361328,
+        size.height * 8.464844,
+        size.width * -1.066406,
+        size.height * 8.425781);
+    path_0.cubicTo(
+        size.width * -1.552734,
+        size.height * 10.50000,
+        size.width * -1.066406,
+        size.height * 11.47266,
+        size.width * -0.4511719,
+        size.height * 9.652344);
+    path_0.cubicTo(
+        size.width * 1.513672,
+        size.height * 11.62500,
+        size.width * 2.871094,
+        size.height * 11.48828,
+        size.width * NaN,
+        size.height * NaN);
+    path_0.close();
+
+    final Paint paint0Fill = Paint()..style = PaintingStyle.fill;
+    paint0Fill.color = const Color(0xff000000).withOpacity(1);
+    canvas.drawPath(path_0, paint0Fill);
+
+    final Path path_1 = Path();
+    path_1.moveTo(size.width * 3.199219, size.height * 6.802734);
+    path_1.cubicTo(
+        size.width * 2.884766,
+        size.height * 6.488281,
+        size.width * 2.832031,
+        size.height * 6.427734,
+        size.width * 2.832031,
+        size.height * 6.376953);
+    path_1.cubicTo(
+        size.width * 2.832031,
+        size.height * 6.326172,
+        size.width * 2.916016,
+        size.height * 6.234375,
+        size.width * 3.490234,
+        size.height * 5.660156);
+    path_1.lineTo(size.width * 4.150391, size.height * 5.000000);
+    path_1.lineTo(size.width * 3.490234, size.height * 4.339844);
+    path_1.cubicTo(
+        size.width * 2.921875,
+        size.height * 3.771484,
+        size.width * 2.832031,
+        size.height * 3.673828,
+        size.width * 2.832031,
+        size.height * 3.623047);
+    path_1.cubicTo(
+        size.width * 2.832031,
+        size.height * 3.574219,
+        size.width * 2.888672,
+        size.height * 3.509766,
+        size.width * 3.199219,
+        size.height * 3.199219);
+    path_1.cubicTo(
+        size.width * 2.867188,
+        size.height * 3.572266,
+        size.width * 2.832031,
+        size.height * 3.630859,
+        size.width * 2.832031,
+        size.height * 3.259766);
+    path_1.cubicTo(
+        size.width * 2.832031,
+        size.height * 3.380859,
+        size.width * 2.886719,
+        size.height * 3.982422,
+        size.width * 3.486328,
+        size.height * NaN);
+    path_1.lineTo(size.width * 4.140625, size.height * NaN);
+    path_1.lineTo(size.width * 4.796875, size.height * NaN);
+    path_1.cubicTo(
+        size.width * 5.361328,
+        size.height * NaN,
+        size.width * 4.142578,
+        size.height * NaN,
+        size.width * 4.142578,
+        size.height * NaN);
+    path_1.cubicTo(
+        size.width * 4.142578,
+        size.height * NaN,
+        size.width * 4.199219,
+        size.height * NaN,
+        size.width * 4.509766,
+        size.height * NaN);
+    path_1.cubicTo(
+        size.width * 4.843750,
+        size.height * NaN,
+        size.width * 4.882813,
+        size.height * NaN,
+        size.width * 4.943359,
+        size.height * NaN);
+    path_1.cubicTo(
+        size.width * 5.005859,
+        size.height * NaN,
+        size.width * 5.056641,
+        size.height * NaN,
+        size.width * 5.654297,
+        size.height * NaN);
+    path_1.cubicTo(
+        size.width * 6.007813,
+        size.height * NaN,
+        size.width * 6.302734,
+        size.height * NaN,
+        size.width * 6.310547,
+        size.height * NaN);
+    path_1.cubicTo(
+        size.width * 6.318359,
+        size.height * NaN,
+        size.width * 6.615234,
+        size.height * NaN,
+        size.width * 6.968750,
+        size.height * NaN);
+    path_1.cubicTo(
+        size.width * 7.562500,
+        size.height * NaN,
+        size.width * 7.615234,
+        size.height * NaN,
+        size.width * 6.968750,
+        size.height * NaN);
+    path_1.cubicTo(
+        size.width * 6.935547,
+        size.height * NaN,
+        size.width * 6.601563,
+        size.height * NaN,
+        size.width * 6.658203,
+        size.height * NaN);
+    path_1.cubicTo(
+        size.width * 6.283203,
+        size.height * NaN,
+        size.width * 6.234375,
+        size.height * NaN,
+        size.width * 6.607422,
+        size.height * NaN);
+    path_1.cubicTo(
+        size.width * 6.458984,
+        size.height * NaN,
+        size.width * 5.892578,
+        size.height * NaN,
+        size.width * NaN,
+        size.height * NaN);
+    path_1.lineTo(size.width * NaN, size.height * NaN);
+    path_1.lineTo(size.width * NaN, size.height * NaN);
+    path_1.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_1.cubicTo(size.width * NaN, size.height * NaN, size.width * NaN,
+        size.height * NaN, size.width * NaN, size.height * NaN);
+    path_1.close();
+
+    final Paint paint1Fill = Paint()..style = PaintingStyle.fill;
+    paint1Fill.color = const Color(0xff000000).withOpacity(1);
+    canvas.drawPath(path_1, paint1Fill);
+  }
+
+  @override
+  bool shouldRepaint(covariant CustomPainter oldDelegate) {
+    return true;
   }
 }